feat: unify missing operator roles functionality
